Tests/LibWeb/Text/expected/element-scrollby-event.txt

scroll event fired y=151 x=25
